Fix table refresh after schedule suppression
