+++ .editorconfig
@@ -17,3 +17,7 @@ indent_size = 2
 
 [Makefile]
 indent_style = tab
+
+[*.{ts, tsx}]
+indent_style = space
+indent_size = 2
