+++ app/src/panels/SchedulePanel.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {useState} from 'react';
 import Box from '@mui/material/Box';
 import FormControl from '@mui/material/FormControl';
-import ScheduleDataTable from "./ScheduleDataTable";
+import ScheduleDataTable from "../components/ScheduleDataTable";
 import {Fab, Snackbar} from "@mui/material";
 import AddIcon from '@mui/icons-material/Add';
 import AlertContext, {IAlert} from "../contexts/Alert";
